Reset Seek Rate
This adds  methods to reset the seek rate back to the default seek rate defined in the config file.
A config variable was defined named home_resets_seek_rate.   This defaults to false if it does not exist in the config file.  If set to true then home will reset the seek rate back to the default seek rate defined in the config file.   This affects G28.2 (or G28 in GRBL mode) Move to preset position will also reset the seek rate.

If move_to_origin_after_home is true and home_resets_seek_rate is true then the move to origin after home will use the default seek rate in config

G28.5 was added to reset the seek rate back to default without actually homing anything